Added config option 'contextSnippetsPerResult'.
This specifies how many context snippets should be included 
per search result when using the OpenSearch API. A context
snippet is a small extract of text from the source document
containing an instance of a search term highlighted with <b>
</b> tags. The snippets are concatenated separated by ellipses
and returned in the <description> tag in each record in the XML
results returned to the client.

Previously this value was hard-coded to a value of 8 in the source
code which produced too much context. 

I have set the default value to 2 instead in web.xml